src/api/requirements.txt
fastapi==0.95.2
fastapi-jwt-auth==0.5.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
uvicorn==0.22.0
black==23.3.0
ruff==0.0.272
SQLAlchemy==2.0.21
alembic==1.12.0
python-jose==3.3.0
passlib==1.7.4
